Add files via upload
These files represent the work I have done in python throughout the duration of the course, Python in Business. This was the second course I have taken in python and learned a substantial amount about how to code. I am posting this a representation of the foundation of my knowledge in the language.